requirements.txt
dlib==19.22.99
firebase_admin==5.2.0
Flask==2.1.1
imutils==0.5.4
matplotlib==3.5.1
numpy==1.22.3
opencv_python==4.5.5.64
Pillow==9.1.0
protobuf==3.20.1
pyrebase==3.0.27
pyrebase5==5.0.1
torch==1.11.0
torchvision==0.12.0
Werkzeug==2.1.0

opencv-python~=4.5.5.64
keras~=2.8.0
tensorflow~=2.8.0